+++ numpoly/array_function/greater.py
@@ -70,7 +70,7 @@ def greater(x1, x2, out=None, **kwargs):
         indices = (coefficients1[idx] != 0) | (coefficients2[idx] != 0)
         indices &= coefficients1[idx] != coefficients2[idx]
         out[indices] = numpy.greater(
-            coefficients1[idx][indices], coefficients2[idx][indices], **kwargs)
+            coefficients1[idx], coefficients2[idx], **kwargs)[indices]
 
     if not out.size:
         out = bool(out)