+++ numpoly/array_function/less.py
@@ -70,6 +70,6 @@ def less(x1, x2, out=None, **kwargs):
         indices = (coefficients1[idx] != 0) | (coefficients2[idx] != 0)
         indices &= coefficients1[idx] != coefficients2[idx]
         out[indices] = numpy.less(
-            coefficients1[idx][indices], coefficients2[idx][indices], **kwargs)
+            coefficients1[idx], coefficients2[idx], **kwargs)[indices]
 
     return out